Feat. add useMutateLogout
